+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Raycast Departures Changelog
 
-## [Initial Version] - 2023-05-05
+## [Added skjer.men integration] - 2023-05-11
+
+- Added action to open stop place on skjer.men
+
+## [Added Departures] - 2023-05-11
+
+Initial version